Ensure celery uses the correct backend set in conf
In fact, it's appear that celery use a decorator @cached_property
that cache the created result backend set, which is
celery.backends.base.DisabledBackend
This depends of wich import you are doing before pyramid
instanciate the Configurator class.
So to ensure the conf is reloaded, cached_property have to be
deleted.

Also if tasks has been registred, ensure that the backend is used
for those tasks.

Looks like #4